+++ .prettierrc.yaml
@@ -1,0 +1,3 @@
+tabWidth: 4
+semi: true
+singleQuote: true
