+++ eslint.config.mjs
@@ -1,0 +1,13 @@
+import js from "@eslint/js";
+import globals from "globals";
+import pluginVue from "eslint-plugin-vue";
+import { defineConfig } from "eslint/config";
+import config from "eslint-config-prettier";
+import plugin from "eslint-plugin-prettier/recommended";
+
+export default defineConfig([
+  { files: ["**/*.{js,mjs,cjs,vue}"], plugins: { js }, extends: ["js/recommended"], languageOptions: { globals: globals.browser } },
+  pluginVue.configs["flat/essential"],
+  config,
+  plugin
+]);
