+++ index.html
@@ -9,9 +9,8 @@
 
 <body>
   <div class="container">
-    <ul class="comments">
-      
-    </ul>
+    <ul class="comments"></ul>
+
     <div class="add-form">
       <input type="text" class="add-form-name" placeholder="Введите ваше имя" />
       <textarea type="textarea" class="add-form-text" placeholder="Введите ваш коментарий" rows="4"></textarea>
@@ -19,9 +18,10 @@
         <button class="add-form-button">Написать</button>
       </div>
     </div>
+
   </div>
 
-  <script type="module" src="index.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 
 </html>
