+++ package.json
@@ -17,5 +17,12 @@
   "bugs": {
     "url": "https://github.com/osilver17/webdev-dom-homework/issues"
   },
-  "homepage": "https://github.com/osilver17/webdev-dom-homework#readme"
+  "homepage": "https://github.com/osilver17/webdev-dom-homework#readme",
+  "devDependencies": {
+    "@eslint/js": "^9.39.2",
+    "eslint": "^4.1.1",
+    "eslint-plugin-vue": "^4.7.1",
+    "globals": "^17.3.0",
+    "prettier": "^3.8.1"
+  }
 }
